- refactor code to separate logic and reduce code complexity
