+++ server/src/app.module.ts
@@ -1,25 +1,38 @@
-import { Module } from '@nestjs/common';
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { HttpConfigModule } from './shared/http-config-module/http-config.module';
+import { NewsController } from './controllers/news/news.controller';
+import { StreetcodeController } from './controllers/streetcode/streetcode.controller';
+import { APP_FILTER } from '@nestjs/core';
+import { ApiExceptionFilter } from './shared/api-exeption-filter/api-exeption.filter';
+import { FileMiddleware } from './shared/file-middleware/file.middleware';
+import { GetAppService } from './shared/get-app-service/get-app.service';
+import { NewsService } from './controllers/news/news.service';
+import { StreetcodeService } from './controllers/streetcode/streetcode.service';
 import { ClientController } from './controllers/client/client.controller';
-import { ClientService } from './controllers/client/client.service';
-import { ClientModule } from './controllers/client/client.module';
-import { NewsService } from './controllers/client/services/news/news.service';
-import { NewsModule } from './controllers/client/services/news/news.module';
-import { StreetcodeService } from './controllers/client/services/streetcodes/streetcode.service';
-import { StreetcodeModule } from './controllers/client/services/streetcodes/streetcode.module';
-import { HttpConfigModule } from './shared/http-config/http-config.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
-    ClientModule,
-    NewsModule,
-    StreetcodeModule,
     HttpConfigModule,
   ],
-  controllers: [ClientController],
-  providers: [ClientService, NewsService, StreetcodeService],
+  controllers: [StreetcodeController, NewsController, ClientController],
+  providers: [
+    NewsService,
+    StreetcodeService,
+    GetAppService,
+    {
+      provide: APP_FILTER,
+      useClass: ApiExceptionFilter,
+    },
+  ],
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer
+      .apply(FileMiddleware)
+      .forRoutes(NewsController, StreetcodeController, ClientController);
+  }
+}
