+++ server/src/controllers/client/client.module.ts
@@ -1,29 +1,7 @@
-import { MiddlewareConsumer, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { ClientController } from './client.controller';
-import { ClientService } from './client.service';
-import { NewsService } from './services/news/news.service';
-import { NewsModule } from './services/news/news.module';
-import { StreetcodeService } from './services/streetcodes/streetcode.service';
-import { HttpConfigModule } from '../../shared/http-config/http-config.module';
-import { ApiExceptionFilter } from '../../shared/api-exeption/api-exeption.filter';
-import { APP_FILTER } from '@nestjs/core';
-import { FileMiddleware } from '../../shared/file-middleware/file.middleware';
 
 @Module({
-  imports: [HttpConfigModule, NewsModule],
-  providers: [
-    NewsService,
-    ClientService,
-    StreetcodeService,
-    {
-      provide: APP_FILTER,
-      useClass: ApiExceptionFilter,
-    },
-  ],
   controllers: [ClientController],
 })
-export class ClientModule {
-  configure(consumer: MiddlewareConsumer) {
-    consumer.apply(FileMiddleware).forRoutes(ClientController);
-  }
-}
+export class ClientModule {}
