--- server/src/controllers/client/services/news/news.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { NewsService } from './news.service';
-import { HttpConfigModule } from '../../../../shared/http-config/http-config.module';
-
-@Module({
-  imports: [HttpConfigModule],
-  providers: [NewsService],
-})
-export class NewsModule {}
