+++ src/features/StreetcodePage/MapBlock/Map/Marker/MarkerWrapper.component.tsx
@@ -4,7 +4,7 @@ import { Marker, Popup } from "react-leaflet";
 import L from 'leaflet';
 
 interface Props {
-    latitude: number, 
+    latitude: number,
     longtitude: number,
     description: string,
     title: string,
@@ -12,7 +12,7 @@ interface Props {
 }
 
 const CustomMarker = ({ description, latitude, longtitude, isStreetcode }: Props) => {
-    const markerClassName = isStreetcode ? "markerCircle streetcodes" : "markerCircle";
+    const markerClassName = isStreetcode ? "markerCircle streetcode" : "markerCircle";
 
     const myIcon = L.divIcon({
         html: `<div class="${markerClassName}"></div>`,
